+++ packages/app/features/about/lib/src/ui/about_page.dart
@@ -5,9 +5,9 @@ import 'package:app_cores_designsystem/common_assets.dart';
 import 'package:app_cores_designsystem/ui.dart';
 import 'package:app_cores_settings/ui.dart';
 import 'package:app_features_about/l10n.dart';
+import 'package:app_features_about/src/ui/contributors/contributors_page.dart';
 import 'package:app_features_about/src/ui/map/map_item_widget.dart';
 import 'package:app_features_about/src/ui/sponsors/sponsors_page.dart';
-import 'package:app_features_about/src/ui/staff/contributors_page.dart';
 import 'package:app_features_about/src/ui/staff/staff_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
